+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+* Breaking Changes
+  * DropInRequest
+    * Remove `collectDeviceData` from `DropInRequest` in favor of always returning device data on `DropInResult`
+
 ## 6.0.0-beta1
 
 * Dependencies